+++ app/src/main/java/io/chthonic/igdb/poc/business/service/IgdbService.kt
@@ -1,12 +1,14 @@
 package io.chthonic.igdb.poc.business.service
 
 import android.content.Context
+import android.support.v4.util.LruCache
 import io.chthonic.igdb.poc.data.client.RestClient
 import io.chthonic.igdb.poc.data.model.IgdbGame
 import io.chthonic.igdb.poc.data.model.IgdbImage
 import io.chthonic.igdb.poc.data.rest.IgdbApi
 import io.chthonic.igdb.poc.data.rest.IgdbApicalypseMapping
 import io.reactivex.Single
+import timber.log.Timber
 
 /**
  * Created by jhavatar on 3/25/2018.
@@ -17,6 +19,10 @@ class IgdbService(appContext: Context) {
         private const val BASE_URL = "https://api-v3.igdb.com/"
         private val DEFAULT_HEADERS = mapOf<String, String>(Pair("user-key", "e35f126c2ae8a127b7696628d0d38260"))
         const val PAGE_SIZE = 20
+    }
+
+    private val coverImageCache by lazy {
+        LruCache<Long, IgdbImage>(500)
     }
 
     private val restClient: RestClient by lazy {
@@ -45,5 +51,51 @@ class IgdbService(appContext: Context) {
 
     fun fetchCoverImages(idList: List<Long>): Single<List<IgdbImage>> {
         return igdbApicalypse.getCoverImages(idList)
+    }
+
+    fun fetchCoverImagesOptimized(gameList: List<IgdbGame>): Single<Map<Long, IgdbImage>> {
+        val coverIds: List<Long> = gameList.mapNotNull {
+            it.cover
+        }
+        Timber.d("coverIds = $coverIds")
+        val cachedImagesMap = coverIds.mapNotNull {
+            val img = coverImageCache.get(it)
+            if (img != null) {
+                coverImageCache.put(it, img) // re-cache
+                it to img
+
+            } else {
+                null
+            }
+
+        }.toMap()
+        Timber.d("cachedImagesMap size = ${cachedImagesMap.size}")
+
+        return if (cachedImagesMap.size == coverIds.size) {
+            Single.fromCallable{
+                cachedImagesMap.toMap()
+            }
+
+        } else {
+
+            val missingCoverIds = coverIds.filter {
+                !cachedImagesMap.containsKey(it)
+            }
+            Timber.d("missingCoverIds = $missingCoverIds")
+
+            fetchCoverImages(missingCoverIds).map { coverList ->
+
+                val fetchedImagesMap = coverList.map {
+                    coverImageCache.put(it.id, it) // cache
+                    it.id to it
+                }.toMap()
+                Timber.d("fetchedImagesMap size = ${fetchedImagesMap.size}")
+
+                val imagesMap = fetchedImagesMap.toMutableMap()
+                imagesMap.putAll(cachedImagesMap)
+
+                imagesMap.toMap()
+            }
+        }
     }
 }
